test: add TSV integration harness, slow test split, and CI
- add TSV-driven CLI integration test harness (integration/run.sh)
- support stdin, exit code, stdout/stderr, and match modes (contains/exact)
- ensure safe backup/restore of profiles and config during runs

- split slow tests into *.slow.test.js
- add npm scripts:
  - test (fast + golden)
  - test:slow (slow tests only)
  - test:all (full suite)

- add GitHub Actions workflow to run full test suite and integration harness on push/PR

no changes to core functionality
