+++ .github/workflows/ci.yml
@@ -1,0 +1,31 @@
+name: CI
+
+on:
+  push:
+  pull_request:
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: npm
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Run full Vitest suite
+        run: npm run test:all
+
+      - name: Make integration runner executable
+        run: chmod +x integration/run.sh
+
+      - name: Run TSV integration harness
+        run: ./integration/run.sh
