+++ integration/config.json
@@ -1,0 +1,4 @@
+{
+  "timeout": 90,
+  "sortBy": "label"
+}
