+++ test/kdf.test.js
@@ -63,14 +63,3 @@ describe('deriveSiteKey', () => {
 
 })
 
-describe('deriveSiteKey (production Argon2)', () => {
-  it('works with production Argon2 parameters [slow]', async () => {
-    const context = encodeContext(DEFAULT_PROFILE)
-
-    const key = await deriveSiteKey('master', context)
-
-    expect(key).toBeInstanceOf(Uint8Array)
-    expect(key).toHaveLength(32)
-  })
-
-})
